Porting redux on top of new key values as is. Need more time to think about changes to redux and meanwhile there's value in moving to more solid key value foundation already
Removing legacy code, keeping the bare minimum required for migration
