--- extensions.go
@@ -1,8 +1,0 @@
-package kevlar
-
-const (
-	JsonExt = ".json"
-	GobExt  = ".gob"
-	HtmlExt = ".html"
-	XmlExt  = ".xml"
-)
