+++ go.mod
@@ -6,7 +6,5 @@ require (
 	github.com/boggydigital/busan v0.1.0
 	github.com/boggydigital/testo v0.1.8
 	github.com/boggydigital/wits v0.2.3
-	golang.org/x/exp v0.0.0-20240823005443-9b4947da3948
+	golang.org/x/exp v0.0.0-20250128182459-e0ece0dbea4c
 )
-
-require github.com/boggydigital/nod v0.1.21
