--- kevlar_legacy/log_records.go
@@ -1,9 +1,0 @@
-package kevlar_legacy
-
-type LogRecord struct {
-	Ts int64
-	Mt MutationType
-	Id string
-}
-
-type LogRecords []*LogRecord
