--- kevlar_legacy/match_options.go
@@ -1,8 +1,0 @@
-package kevlar_legacy
-
-type MatchOption int
-
-const (
-	CaseSensitive = iota
-	FullMatch
-)
