--- kevlar_legacy/mutation_types.go
@@ -1,9 +1,0 @@
-package kevlar_legacy
-
-type MutationType int
-
-const (
-	Create MutationType = iota
-	Update
-	Cut
-)
