+++ redux_reader.go
@@ -1,4 +1,4 @@
-package kevlar_legacy
+package kevlar
 
 import (
 	"golang.org/x/exp/maps"
