+++ redux_reader_match_test.go
@@ -1,4 +1,4 @@
-package kevlar_legacy
+package kevlar
 
 import (
 	"github.com/boggydigital/testo"
