+++ redux_reader_sort.go
@@ -1,4 +1,4 @@
-package kevlar_legacy
+package kevlar
 
 import "sort"
